Migrate old website patches to Hugo overrides
Instead of directly modifying the theme files, use Hugo's override feature. All
site-specific changes live in the top-level directories. When Hugo generates the
static site, it merges the top-level files with the theme files, adding any new
files, and replacing any theme files with the same name and path as the
top-level files. This clean separation between the theme and local changes makes
it easier to upgrade to future releases of the theme.